Module Reservoirs (MOHID Land) update
Module Reservoirs (MOHID Land) : Corrected bug to allow more than one
reservoir and added option to be able to do a continuous simulation
Module Reservoirs, Module Drainage Network and Module Basin : changes to
allow compilation in single precision

+++ Software/MOHIDLand/ModuleBasin.F90
@@ -5644,7 +5644,7 @@ cd2 :           if (BlockFound) then
         real, dimension(:, :   ), pointer           :: RPConcentration
         real, dimension(:      ), pointer           :: ReservoirConcentration
         integer, dimension(:, :), pointer           :: GWLayer
-        real(8), dimension(:), pointer              :: ReservoirFlowToChannels
+        real, dimension(:), pointer                 :: ReservoirFlowToChannels
 
         if (MonitorPerformance) call StartWatch ("ModuleBasin", "DrainageNetworkProcesses")
         
@@ -6001,7 +6001,7 @@ cd2 :           if (BlockFound) then
         integer                                     :: nProperties, iProp 
         integer                                     :: PropID
         real, dimension(:   ), pointer              :: DNConcentration
-        real(8), dimension(:), pointer              :: ChannelFlowToReservoir
+        real, dimension(:), pointer                 :: ChannelFlowToReservoir
         logical                                     :: PropAdvDiff
 
         if (MonitorPerformance) call StartWatch ("ModuleBasin", "ReservoirProcesses")
